認証のフロントのリファクタリング
