+++ client/public/docs/openapi.json
@@ -811,21 +811,6 @@
           }
         }
       }
-    },
-    "/shop": {
-      "get": {
-        "responses": {
-          "2XX": {
-            "content": {
-              "application/json": {
-                "schema": {
-                  "type": "string"
-                }
-              }
-            }
-          }
-        }
-      }
     }
   },
   "components": {
